pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.0.0",
    "sqlparse>=0.5.5",
    "tzdata>=2025.3",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "werkzeug>=3.1.4",
]
